文章详情页面、评论展示

+++ src/main/java/learn/qzy/searchbackend/model/entity/ArticleComment.java
@@ -34,7 +34,7 @@ public class ArticleComment implements Serializable {
      * 楼层号
      */
     @TableField(value = "serial_number")
-    private Long serialNumber;
+    private Integer serialNumber;
 
     /**
      * 父级评论用户名
